+++ test.js
@@ -1,9 +1,7 @@
 var Jpex = require('./jpex');
-var Class = Jpex.extend({
-  dependencies : ['$error'],
-  constructor : function($error){
-
-  }
+var Class = Jpex.extend(function($resolve){
+  var $fs = $resolve('$fs');
+  var foo = $resolve('foo');
 });
-Class.Register.ErrorType('DemoError');
-new Class();
+Class.Register.Factory('foo', () => 'bah');
+Class();